refactor: extract useNodeLoadHandlers hook and fix prop types
- Extract node load handler setup (structure, trajectory, vector) from
  MeganeViewer into a new useNodeLoadHandlers custom hook, reducing the
  component from 327 to 254 lines
- Fix MeganeViewerProps: change onBondSourceChange, onLabelSourceChange,
  onVectorSourceChange from (source: string) to proper typed union types
  (BondSource, LabelSource, VectorSource)
- Remove unsafe `as (s: string) => void` casts in index.tsx
